refactor: rename navbar.tsx to index.tsx
